--- a/airflow/utils/dag_processing.py
+++ b/airflow/utils/dag_processing.py
@@ -252,6 +252,14 @@ class AbstractDagFileProcessorProcess(metaclass=ABCMeta):
         """
         raise NotImplementedError()
 
+    @property
+    @abstractmethod
+    def waitable_handle(self):
+        """
+        A "waitable" handle that can be passed to ``multiprocessing.connection.wait()``
+        """
+        raise NotImplementedError()
+
 
 class DagParsingStat(NamedTuple):
     """Information on processing progress"""
@@ -652,6 +660,8 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
 
         self._log = logging.getLogger('airflow.processor_manager')
 
+        self.waitables = {self._signal_conn: self._signal_conn}
+
     def register_exit_signals(self):
         """
         Register signals that stop child processes
@@ -698,13 +708,20 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
         else:
             poll_time = None
 
-        # Used to track how long it takes us to get once around every file in the DAG folder.
-        self._parsing_start_time = timezone.utcnow()
+        self._refresh_dag_dir()
+        self.prepare_file_path_queue()
+
+        if self._async_mode:
+            # If we're in async mode, we can start up straight away. If we're
+            # in sync mode we need to be told to start a "loop"
+            self.start_new_processes()
+
         while True:
             loop_start_time = time.time()
 
             # pylint: disable=no-else-break
-            if self._signal_conn.poll(poll_time):
+            ready = multiprocessing.connection.wait(self.waitables.keys(), timeout=poll_time)
+            if self._signal_conn in ready:
                 agent_signal = self._signal_conn.recv()
                 self.log.debug("Received %s signal from DagFileProcessorAgent", agent_signal)
                 if agent_signal == DagParsingSignal.TERMINATE_MANAGER:
@@ -719,13 +736,36 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
                 elif isinstance(agent_signal, FailureCallbackRequest):
                     self._add_callback_to_queue(agent_signal)
                 else:
-                    raise AirflowException("Invalid message")
-            elif not self._async_mode:
+                    raise ValueError(f"Invalid message {type(agent_signal)}")
+
+            if not ready and not self._async_mode:
                 # In "sync" mode we don't want to parse the DAGs until we
                 # are told to (as that would open another connection to the
                 # SQLite DB which isn't a good practice
+
+                # This shouldn't happen, as in sync mode poll should block for
+                # ever. Lets be defensive about that.
+                self.log.warning(
+                    "wait() unexpectedly returned nothing ready after infinite timeout (%r)!",
+                    poll_time
+                )
+
                 continue
-            # pylint: enable=no-else-break
+
+            for sentinel in ready:
+                if sentinel is self._signal_conn:
+                    continue
+
+                processor = self.waitables.get(sentinel)
+                if not processor:
+                    continue
+
+                simple_dags = self._collect_results_from_processor(processor)
+                self.waitables.pop(sentinel)
+                self._processors.pop(processor.file_path)
+                for simple_dag in simple_dags:
+                    self._signal_conn.send(simple_dag)
+
             self._refresh_dag_dir()
             self._find_zombies()  # pylint: disable=no-value-for-parameter
 
@@ -742,10 +782,6 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
             # Update number of loop iteration.
             self._num_run += 1
 
-            simple_dags = self.collect_results()
-            for simple_dag in simple_dags:
-                self._signal_conn.send(simple_dag)
-
             if not self._async_mode:
                 self.log.debug(
                     "Waiting for processors to finish since we're using sqlite")
@@ -755,10 +791,10 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
                 # this type of message
                 self.wait_until_finished()
 
-                # Collect anything else that has finished, but don't kick off any more processors
-                simple_dags = self.collect_results()
-                for simple_dag in simple_dags:
-                    self._signal_conn.send(simple_dag)
+            # Collect anything else that has finished, but don't kick off any more processors
+            simple_dags = self.collect_results()
+            for simple_dag in simple_dags:
+                self._signal_conn.send(simple_dag)
 
             self._print_stat()
 
@@ -1032,6 +1068,31 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
             while not processor.done:
                 time.sleep(0.1)
 
+    def _collect_results_from_processor(self, processor):
+        self.log.debug("Processor for %s finished", processor.file_path)
+        Stats.decr('dag_processing.processes')
+        last_finish_time = timezone.utcnow()
+
+        if processor.result is not None:
+            dags, count_import_errors = processor.result
+        else:
+            self.log.error(
+                "Processor for %s exited with return code %s.",
+                processor.file_path, processor.exit_code
+            )
+            dags, count_import_errors = [], -1
+
+        stat = DagFileStat(
+            num_dags=len(dags),
+            import_errors=count_import_errors,
+            last_finish_time=last_finish_time,
+            last_duration=(last_finish_time - processor.start_time).total_seconds(),
+            run_count=self.get_run_count(processor.file_path) + 1,
+        )
+        self._file_stats[processor.file_path] = stat
+
+        return dags
+
     def collect_results(self):
         """
         Collect the result from any finished DAG processors
@@ -1040,50 +1101,22 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
             have finished since the last time this was called
         :rtype: list[airflow.utils.dag_processing.SimpleDag]
         """
-        finished_processors: Dict[str, AbstractDagFileProcessorProcess] = {}
-        running_processors: Dict[str, AbstractDagFileProcessorProcess] = {}
+        # Collect all the DAGs that were found in the processed files
+        simple_dags = []
 
-        for file_path, processor in self._processors.items():
-            if processor.done:
-                self.log.debug("Processor for %s finished", file_path)
-                Stats.decr('dag_processing.processes')
-                last_finish_time = timezone.utcnow()
-                finished_processors[file_path] = processor
+        ready = multiprocessing.connection.wait(self.waitables.keys() - [self._signal_conn], timeout=0)
 
-                if processor.result is not None:
-                    dags, count_import_errors = processor.result
-                else:
-                    dags, count_import_errors = [], -1
-
-                stat = DagFileStat(
-                    num_dags=len(dags),
-                    import_errors=count_import_errors,
-                    last_finish_time=last_finish_time,
-                    last_duration=(last_finish_time - processor.start_time).total_seconds(),
-                    run_count=self.get_run_count(file_path) + 1,
-                )
-                self._file_stats[file_path] = stat
-            else:
-                running_processors[file_path] = processor
-        self._processors = running_processors
+        for sentinel in ready:
+            processor = self.waitables[sentinel]
+            self.waitables.pop(processor.waitable_handle)
+            self._processors.pop(processor.file_path)
+            simple_dags += self._collect_results_from_processor(processor)
 
         self.log.debug("%s/%s DAG parsing processes running",
                        len(self._processors), self._parallelism)
 
         self.log.debug("%s file paths queued for processing",
                        len(self._file_path_queue))
-
-        # Collect all the DAGs that were found in the processed files
-        simple_dags = []
-        for file_path, processor in finished_processors.items():
-            if processor.result is None:
-                self.log.error(
-                    "Processor for %s exited with return code %s.",
-                    processor.file_path, processor.exit_code
-                )
-            else:
-                for simple_dag in processor.result[0]:
-                    simple_dags.append(simple_dag)
 
         return simple_dags
 
@@ -1109,6 +1142,7 @@ class DagFileProcessorManager(LoggingMixin):  # pylint: disable=too-many-instanc
                 processor.pid, file_path
             )
             self._processors[file_path] = processor
+            self.waitables[processor.waitable_handle] = processor
 
     def prepare_file_path_queue(self):
         """