@ashb 2020-05-11
I don't think this will affect many real workloads, but does affect the "I'm just playing around with Airflow" and seeing it being slow. There's still a lot of work to do there, but this is an improvement.