@ashb 2020-05-11
This (as written) only works on Py3, so we can't backport it, not without checking for `.sentinel` attributes.