@ashb 2020-05-11
I think I've got some infinite loop being triggered on CI that wasn't happening locally. Hmm.