@ashb 2020-05-12
Running locally:

```
================================================ 5594 passed, 286 skipped, 2 xfailed, 1 xpassed, 477 warnings in 969.54s (0:16:09) ================================================```

HRMMMM.