@ashb 2020-05-12
Ah okay, I've managed to reproduce it, but only when running the whole test suite (It helps if you run the right tests when you have the right branch checked out!)