@ashb 2020-05-15
Think I've finally worked out why the test_dags_system_exit was _sometimes_ failing now. I started the processes earlier so it's possible that it will have run and had results collected twice as a result.