@ashb 2020-05-15
Double checked before+after timing, still valid.